requirements.txt
scikit-learn
pandas
numpy
mysql-connector-python
matplotlib